New use case robo electric
