+++ app/src/test/java/nl/tudelft/cs4160/trustchain_android/ExampleInstrumentedTest.java
@@ -1,0 +1,26 @@
+//package nl.tudelft.cs4160.trustchain_android;
+//
+//import android.content.Context;
+//import android.support.test.InstrumentationRegistry;
+//import android.support.test.runner.AndroidJUnit4;
+//
+//import org.junit.Test;
+//import org.junit.runner.RunWith;
+//
+//import static org.junit.Assert.assertEquals;
+//
+///**
+// * Instrumentation test, which will execute on an Android device.
+// *
+// * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+// */
+//@RunWith(AndroidJUnit4.class)
+//public class ExampleInstrumentedTest {
+//    @Test
+//    public void useAppContext() throws Exception {
+//        // Context of the app under test.
+//        Context appContext = InstrumentationRegistry.getTargetContext();
+//
+//        assertEquals("nl.tudelft.cs4160.trustchain_android", appContext.getPackageName());
+//    }
+//}
