+++ app/src/test/java/nl/tudelft/cs4160/trustchain_android/chainExplorer/ChainExplorerAdapterTest.java
