+++ codecov.yml
@@ -3,7 +3,6 @@
 #########################################
 coverage:
   ignore:
-  - "app/src/main/java/nl/tudelft/cs4160/trustchain_android/main/.*"
   - "app/src/main/java/nl/tudelft/cs4160/trustchain_android/database/.*"
   - "app/src/main/java/nl/tudelft/cs4160/trustchain_android/KeyActivity.java"
   - "app/src/main/java/nl/tudelft/cs4160/trustchain_android/message/.*"
